Bug with calculation feat slots has been fixed.
