+++ src/cmd_god/show.cpp
@@ -7,6 +7,7 @@
 
 #include "ban.h"
 #include "administration/privilege.h"
+#include "cmd/do_features.h"
 #include "cmd_god/shutdown_parameters.h"
 #include "communication/parcel.h"
 #include "communication/mail.h"
@@ -35,7 +36,6 @@ extern const char *Dirs[];
 extern bool is_empty(ZoneRnum zone_nr);
 extern void show_apply(CharData *ch, CharData *vict);
 extern void print_rune_stats(CharData *ch);
-extern void list_feats(CharData *ch, CharData *vict, bool all_feats);
 extern void list_skills(CharData *ch, CharData *vict, const char *filter = nullptr);
 extern void list_spells(CharData *ch, CharData *vict, int all_spells);
 
@@ -636,7 +636,7 @@ void do_show(CharData *ch, char *argument, int/* cmd*/, int/* subcmd*/) {
 				SendMsgToChar("Нет такого игрока.\r\n", ch);
 				return;
 			}
-			list_feats(vict, ch, false);
+			DisplayFeats(vict, ch, false);
 			break;
 		case 17:        // show glory
 			GloryMisc::show_log(ch, value);
